Fixed some type conversion warnings

+++ src/common/pa_converters.c
@@ -341,7 +341,7 @@ static void Float32_To_Int32(
     while( count-- )
     {
         /* REVIEW */
-        double scaled = *src * 0x7FFFFFFF;
+        double scaled = *src * (double)0x7FFFFFFF;
         *dest = (PaInt32) scaled;
 
         src += sourceStride;
@@ -386,7 +386,7 @@ static void Float32_To_Int32_Clip(
     while( count-- )
     {
         /* REVIEW */
-        double scaled = *src * 0x7FFFFFFF;
+        double scaled = *src * (double)0x7FFFFFFF;
         PA_CLIP_( scaled, -2147483648., 2147483647.  );
         *dest = (PaInt32) scaled;
 
@@ -505,7 +505,7 @@ static void Float32_To_Int24_Clip(
     while( count-- )
     {
         /* convert to 32 bit and drop the low 8 bits */
-        double scaled = *src * 0x7FFFFFFF;
+        double scaled = *src * (double)0x7FFFFFFF;
         PA_CLIP_( scaled, -2147483648., 2147483647.  );
         temp = (PaInt32) scaled;
